pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alienx"
version = "1.0.0"
description = "Voice-controlled PC automation assistant"
readme = "README.md"
authors = [{name = "Mohd Mujtaba", email = "mujtaba.aamir1@gmail.com"}]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: POSIX :: Linux",
]
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "pynput>=1.7.6",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
mac = [
    "pyobjc-framework-Cocoa>=10.0",
    "pyobjc-framework-Quartz>=10.0",
]

[project.scripts]
alienx = "src.main:main"

[tool.setuptools]
package-dir = {"" = "src"}
